requirements.txt
# Core web framework
Django==4.2
django-crispy-forms==2.0
whitenoise==6.5.0
reportlab==4.0.4

# ML and data processing
# Current fraud scoring backend: scikit-learn
# Optional future deep learning backends: TensorFlow or PyTorch
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2

# OCR and computer vision
# Note: pytesseract also requires the Tesseract OCR system binary to be installed.
pytesseract==0.3.10
Pillow==10.0.0
opencv-python-headless==4.8.0.76
pdf2image==1.17.0
